🎨 (refactor):Separation of components into 2 different pages
